% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{array_to_cr}
\alias{array_to_cr}
\title{Convert numeric array to cairo raster}
\usage{
array_to_cr(arr, cr = NULL)
}
\arguments{
\item{arr}{Numeric R array}

\item{cr}{optional cairo raster as a preallocation destination. If NULL (default), 
then a new cairo raster will be created}
}
\value{
cairo raster
}
\description{
Convert numeric array to cairo raster
}
\examples{
logo_file <- system.file("img/logo-rainbow.png", package = 'cara', mustWork = TRUE)
logo <- cr_read_png(logo_file)
arr <- cr_to_array(logo)
class(arr)
plot(as.raster(arr))
logo2 <- array_to_cr(arr)
plot(logo2, T)
}
\seealso{
Other image conversion functions: 
\code{\link{cr_to_array}()},
\code{\link{cr_to_nr}()},
\code{\link{cr_to_raster}()},
\code{\link{magick_to_cr}()},
\code{\link{matrix_to_cr}()},
\code{\link{nr_to_cr}()},
\code{\link{raster_to_cr}()}
}
\concept{image conversion functions}
