% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-raster.R
\name{cairo raster}
\alias{cairo raster}
\title{Cairo raster}
\description{
Cairo is a graphics processing library used for drawing operations in R 
and many other programs.  Cairo refers to its drawing canvas as an
\emph{image surface} and this canvas is exposed by this package as a
\emph{cairo raster}.
}
\details{
In general, access to the drawing capabilities of Cairo is mediated through
R's devices.  Devices are a useful abstraction which allow R to 
communicate with many different graphics devices through a common interface.

However, R's graphics devices don't expose the drawing canvas for 
low level manipulation.

This package makes it possible to manipulate the pixels with Cairo 
and draw directly upon cairo raster objects (known within Cairo documentation
as image surfaces).

The drawing canvas in C (a Cairo image surface) is wrapped as an R
external pointer.  This pointer is passed to various functions in order
draw upon it, manipulate it and copy it to other images.

The cairo raster can be converted to other R object types (e.g. array, raster)
if you want direct access to the pixel values.
}
