% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-params.R
\name{cpar}
\alias{cpar}
\title{Create a list of graphical parameters for cairo}
\usage{
cpar(
  antialias = NULL,
  line_width = NULL,
  line_end = NULL,
  line_join = NULL,
  line_mitre_limit = NULL,
  line_dash = NULL,
  operator = NULL,
  fill_rule = NULL
)
}
\arguments{
\item{antialias}{antialias setting. One of: 
c('default', none', 'gray', 'subpixel', 'fast', 'good', 'best')}

\item{line_width}{Numeric value for width of line. Default: NULL (Use 2.0)}

\item{line_end}{String indicating how line ends should be drawn.
Default: NULL (use 'butt'). Valid values: 'butt', 'round', 'square'
\describe{
   \item{\code{butt}}{start/stop the line exactly at the given coordinates}
   \item{\code{round}}{use a round ending}
   \item{\code{square}}{use a squared ending}
}}

\item{line_join}{String indicating how lines should be joined.
Default: NULL (use 'miter'). Valid values: 'miter', 'round', 'bevel'
\describe{
  \item{\code{miter}}{Use a sharp angled corner}
  \item{\code{round}}{Use a rounded join}
  \item{\code{bevel}}{use a cut-off join}
}}

\item{line_mitre_limit}{Numeric value. If current \code{line_join} is 'miter', then this limit is used
to determine whether the lines should be joined with a bevel instead of a miter.
Default: NULL (use internal default of \code{line_join = 10})}

\item{line_dash}{a vector of positive numeric values given the length of alternate
'on' and 'off' portions of the stroke.}

\item{operator}{String specifying compositing operator. Default: "over".
Valid values: clear, source,
over, in, out, atop, dest, dest-over, dest-in, dest-out, dest-atop,
xor, add, saturate, multiply, screen, overlay, darken, lighten,
dodge, burn, hard-light, soft-light, difference, exclusion, 
hsl-hue, hsl-sat, hsl-color, hsl-luminosity. 
For more details on operators, see Cairo's documentation
\url{https://cairographics.org/operators/}}

\item{fill_rule}{'winding' or 'even-odd'. Default: winding}
}
\value{
Named list of graphical parameters
}
\description{
Create a list of graphical parameters for cairo
}
\examples{
cpar(line_width = 1)
}
\seealso{
Other parameter related functions: 
\code{\link{cr_params_save}()},
\code{\link{cr_params_set}()}
}
\concept{parameter related functions}
