% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-draw.R
\name{cr_circle}
\alias{cr_circle}
\title{Draw circles}
\usage{
cr_circle(cr, x, y, r, fill = "black", color = NA, params = NULL)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{x, y, r}{coordinates}

\item{fill}{fill color. Default: 'black'}

\item{color}{border color. Default: NA (no line)}

\item{params}{graphical parameters created with \code{\link{cpar}()}. Default: NULL
means to use the currently defined global parameters}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Draw circles
}
\examples{
set.seed(3)
img <- cr_new(400, 300)
cr_circle(img, sample(400, 3), sample(300, 3), r= c(10, 50, 100), 
        fill = rainbow(3), color = 'black')
plot(img, T)
}
\seealso{
Other draw functions: 
\code{\link{cr_arc}()},
\code{\link{cr_line}()},
\code{\link{cr_point}()},
\code{\link{cr_polygon}()},
\code{\link{cr_polyline}()},
\code{\link{cr_rect}()},
\code{\link{cr_spline}()}
}
\concept{draw functions}
