% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-core.R
\name{cr_dim}
\alias{cr_dim}
\title{Dimensions of \code{cairo raster}}
\usage{
cr_dim(cr)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}
}
\value{
integer vector of dimensions. i.e. height, width
}
\description{
Dimensions of \code{cairo raster}
}
\examples{
img <- cr_new(120, 50)
cr_dim(img)
}
