% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-core.R
\name{cr_duplicate}
\alias{cr_duplicate}
\title{Create a new \code{cairo raster} image with the contents from an existing image}
\usage{
cr_duplicate(cr)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}
}
\value{
A new \code{cairo raster} object containing a copy of the image from the
        original
}
\description{
Create a new \code{cairo raster} image with the contents from an existing image
}
\examples{
cr1 <- cr_new(200, 200, 'blue')
cr2 <- cr_duplicate(cr1)
plot(cr2, T)

}
