% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-core.R
\name{cr_fill}
\alias{cr_fill}
\title{Fill an image with a color or pattern}
\usage{
cr_fill(cr, fill)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{fill}{color (as hex, colorname or integer) or pattern object 
returned by \code{create_pattern_*} family of functions}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Fill an image with a color or pattern
}
\examples{
img <- cr_new(800, 600, 'hotpink')
plot(img, T)

cr_fill(img, '#0000ff')
plot(img, T)

logo_file <- system.file("img/logo-rainbow.png", package = "cara")
logo <- cr_read_png(logo_file)
logo <- cr_resize(logo, 150, 100)
pat <- create_pattern_image(logo)
cr_fill(img, pat)
plot(img, T)
}
