% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-draw.R
\name{cr_line}
\alias{cr_line}
\title{Draw lines}
\usage{
cr_line(cr, x0, y0, x1, y1, color = "black", params = NULL)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{x0, y0, x1, y1}{coordinates of the endpoints of the lines}

\item{color}{line color. Default: 'black'}

\item{params}{graphical parameters created with \code{\link{cpar}()}. Default: NULL
means to use the currently defined global parameters}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Draw lines
}
\examples{
img <- cr_new(400, 300)
cr_params_set(img, params = cpar(line_dash = c(3, 7)))
cr_line(img, x0 = 0, y0 = c(0, 100, 200), x1 = 400, y1 = c(300, 200, 100), 
        color = rainbow(3), param = cpar(line_width = 5))
plot(img, T)
}
\seealso{
Other draw functions: 
\code{\link{cr_arc}()},
\code{\link{cr_circle}()},
\code{\link{cr_point}()},
\code{\link{cr_polygon}()},
\code{\link{cr_polyline}()},
\code{\link{cr_rect}()},
\code{\link{cr_spline}()}
}
\concept{draw functions}
