% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-mask.R
\name{cr_mask_begin}
\alias{cr_mask_begin}
\alias{cr_mask_end}
\title{Start/end region of masked drawing}
\usage{
cr_mask_begin(cr, mask, x = 0, y = 0)

cr_mask_end(cr)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{mask}{Image (cairo raster) to use as mask.  This image does not need 
to be the same size as \code{cr}}

\item{x, y}{position of mask within the destinatino. Default: (0, 0) i.e. 
start at the origin (top-left)}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
These functions are used to globally set a mask.  Within the bounds of 
these two functions, drawing will only occur where the pixels of the
mask are opaque.
}
\examples{
logo_file <- system.file("img/logo-rainbow.png", package = 'cara', mustWork = TRUE)
logo <- cr_read_png(logo_file)
plot(logo, T)
img <- cr_new(600, 400)

cr_params_set(img, cpar(line_width = 10))

cr_mask_begin(img, mask = logo)
cr_circle(img, 300, 200, 200, fill = 'blue', color = 'black')
cr_circle(img, 350, 200, 100, fill = 'darkgreen', color = 'black')
cr_mask_end(img)

plot(img, T)
}
