% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-params.R
\name{cr_params_save}
\alias{cr_params_save}
\alias{cr_params_restore}
\title{Save and restore (push and pop) global parameter settings on a hidden internal stack}
\usage{
cr_params_save(cr)

cr_params_restore(cr)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Save and restore (push and pop) global parameter settings on a hidden internal stack
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_params_save(cr)
cr_params_restore(cr)
\dontshow{\}) # examplesIf}
}
\seealso{
Other parameter related functions: 
\code{\link{cpar}()},
\code{\link{cr_params_set}()}
}
\concept{parameter related functions}
