% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-params.R
\name{cr_params_set}
\alias{cr_params_set}
\title{Set global drawing parameters.}
\usage{
cr_params_set(cr, params)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{params}{Cairo parameters as generated by \code{\link{cpar}()}}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
These parameters exist globally, but can be overridden by passing a 
\code{params} argument to any drawing function
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_params_set(cr, params = cpar())
\dontshow{\}) # examplesIf}
}
\seealso{
Other parameter related functions: 
\code{\link{cpar}()},
\code{\link{cr_params_save}()}
}
\concept{parameter related functions}
