% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pattern.R
\name{cr_pattern_rotate}
\alias{cr_pattern_rotate}
\title{Rotate}
\usage{
cr_pattern_rotate(pat, angle)
}
\arguments{
\item{pat}{pattern object created by \code{create_pattern_image()} etc}

\item{angle}{rotation angle in radians}
}
\value{
Invisibly return the original \code{cairo pattern} object
}
\description{
Rotate
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_pattern_rotate(pat, pi)
\dontshow{\}) # examplesIf}
}
\seealso{
Other pattern transformation functions: 
\code{\link{cr_pattern_scale}()},
\code{\link{cr_pattern_transform_reset}()},
\code{\link{cr_pattern_translate}()}
}
\concept{pattern transformation functions}
