% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pattern.R
\name{cr_pattern_scale}
\alias{cr_pattern_scale}
\title{Scale}
\usage{
cr_pattern_scale(pat, x, y = x)
}
\arguments{
\item{pat}{pattern object created by \code{create_pattern_image()} etc}

\item{x, y}{scale factors}
}
\value{
Invisibly return the original \code{cairo pattern} object
}
\description{
Scale
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_pattern_scale(pat, 2, 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other pattern transformation functions: 
\code{\link{cr_pattern_rotate}()},
\code{\link{cr_pattern_transform_reset}()},
\code{\link{cr_pattern_translate}()}
}
\concept{pattern transformation functions}
