% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pattern.R
\name{cr_pattern_transform_reset}
\alias{cr_pattern_transform_reset}
\title{Reset transformation}
\usage{
cr_pattern_transform_reset(pat)
}
\arguments{
\item{pat}{pattern object created by \code{create_pattern_image()} etc}
}
\value{
Invisibly return the original \code{cairo pattern} object
}
\description{
Reset transformation
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_pattern_transform_reset(pat)
\dontshow{\}) # examplesIf}
}
\seealso{
Other pattern transformation functions: 
\code{\link{cr_pattern_rotate}()},
\code{\link{cr_pattern_scale}()},
\code{\link{cr_pattern_translate}()}
}
\concept{pattern transformation functions}
