% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pattern.R
\name{cr_pattern_translate}
\alias{cr_pattern_translate}
\title{Translate a pattern}
\usage{
cr_pattern_translate(pat, x, y)
}
\arguments{
\item{pat}{pattern object created by \code{create_pattern_image()} etc}

\item{x, y}{translation offsets}
}
\value{
Invisibly return the original \code{cairo pattern} object
}
\description{
Translate a pattern
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_pattern_translate(pat, 2, 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other pattern transformation functions: 
\code{\link{cr_pattern_rotate}()},
\code{\link{cr_pattern_scale}()},
\code{\link{cr_pattern_transform_reset}()}
}
\concept{pattern transformation functions}
