% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-draw.R
\name{cr_point}
\alias{cr_point}
\title{Draw points (single square pixel)}
\usage{
cr_point(cr, x, y, color = "black", size = 1, params = NULL)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{x, y}{location}

\item{color}{point color. Default: 'black'}

\item{size}{Side length of square representing a point. Default: 1}

\item{params}{graphical parameters created with \code{\link{cpar}()}. Default: NULL
means to use the currently defined global parameters}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Draw points (single square pixel)
}
\examples{
set.seed(3)
img <- cr_new(400, 300, 'grey20')
N <- 1000
cr_point(img, sample(400, N, TRUE), sample(300, N, TRUE), color = rainbow(N))
plot(img, T)
}
\seealso{
Other draw functions: 
\code{\link{cr_arc}()},
\code{\link{cr_circle}()},
\code{\link{cr_line}()},
\code{\link{cr_polygon}()},
\code{\link{cr_polyline}()},
\code{\link{cr_rect}()},
\code{\link{cr_spline}()}
}
\concept{draw functions}
