% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-draw.R
\name{cr_polygon}
\alias{cr_polygon}
\title{Draw polygons}
\usage{
cr_polygon(cr, x, y, fill = "black", color = NA, params = NULL)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{x, y}{vectors of coordinates}

\item{fill}{fill color}

\item{color}{line color. Default: 'black'}

\item{params}{graphical parameters created with \code{\link{cpar}()}. Default: NULL
means to use the currently defined global parameters}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Draw polygons
}
\examples{
w <- h <- 1000
cr <- cr_new(w, h, 'grey80')
set.seed(1)
x <- runif(5, 0, w)
y <- runif(5, 0, h)
cr_params_set(cr, params = cpar(line_dash = c(15, 20), line_width = 4))
cr_polygon(cr, x, y, fill = 'royalblue', color = 'black')
plot(cr, T)
}
\seealso{
Other draw functions: 
\code{\link{cr_arc}()},
\code{\link{cr_circle}()},
\code{\link{cr_line}()},
\code{\link{cr_point}()},
\code{\link{cr_polyline}()},
\code{\link{cr_rect}()},
\code{\link{cr_spline}()}
}
\concept{draw functions}
