% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-draw.R
\name{cr_polyline}
\alias{cr_polyline}
\title{Draw polylines}
\usage{
cr_polyline(cr, x, y, color = "black", params = NULL)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{x, y}{vectors of coordinates}

\item{color}{line color. Default: 'black'}

\item{params}{graphical parameters created with \code{\link{cpar}()}. Default: NULL
means to use the currently defined global parameters}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Draw polylines
}
\examples{
set.seed(3)
w <- h <- 1000
cr <- cr_new(w, h, 'white')
x <- runif(4, 0, w)
y <- runif(4, 0, h)
cr_polyline(cr, x, y, params = cpar(line_width = 20, line_join = 'miter'))
plot(cr, T)
}
\seealso{
Other draw functions: 
\code{\link{cr_arc}()},
\code{\link{cr_circle}()},
\code{\link{cr_line}()},
\code{\link{cr_point}()},
\code{\link{cr_polygon}()},
\code{\link{cr_rect}()},
\code{\link{cr_spline}()}
}
\concept{draw functions}
