% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr-png.R
\name{cr_read_png}
\alias{cr_read_png}
\title{Read PNG as a cairo raster}
\usage{
cr_read_png(filename)
}
\arguments{
\item{filename}{PNG filename}
}
\value{
cairo raster
}
\description{
Read PNG as a cairo raster
}
\examples{
filename <- system.file("img/logo-rainbow.png", package = 'cara', mustWork = TRUE)
cr <- cr_read_png(filename)
plot(cr, T)  
}
