% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-draw.R
\name{cr_rect}
\alias{cr_rect}
\title{Draw rectangles}
\usage{
cr_rect(
  cr,
  x,
  y,
  w,
  h,
  fill = "black",
  color = NA,
  hjust = 0,
  vjust = 0,
  params = NULL
)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{x, y}{location of rectangle}

\item{w, h}{width and height of rectangle}

\item{fill}{fill color. Default: 'black'}

\item{color}{border color. Default: NA (no line)}

\item{hjust, vjust}{justification of the rectangle relative to its \code{(x, y)}
location. Default: (0, 0).  Use (0.5, 0.5) to center the rectangle 
at the (x, y) coordinates}

\item{params}{graphical parameters created with \code{\link{cpar}()}. Default: NULL
means to use the currently defined global parameters}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Draw rectangles
}
\examples{
set.seed(3)
img <- cr_new(400, 300)
cr_rect(img, sample(400, 3), sample(300, 3), w = c(10, 50, 100), h = 42, 
        fill = rainbow(3), color = 'black', params = cpar(line_width = 3))
plot(img, T)
}
\seealso{
Other draw functions: 
\code{\link{cr_arc}()},
\code{\link{cr_circle}()},
\code{\link{cr_line}()},
\code{\link{cr_point}()},
\code{\link{cr_polygon}()},
\code{\link{cr_polyline}()},
\code{\link{cr_spline}()}
}
\concept{draw functions}
