% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-core.R
\name{cr_resize}
\alias{cr_resize}
\title{Resize current image into a new image}
\usage{
cr_resize(cr, width, height)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{width, height}{new dimensions}
}
\value{
New cairo raster object
}
\description{
Resize current image into a new image
}
\examples{
logo_file <- system.file("img/logo-rainbow.png", package = "cara")
logo <- cr_read_png(logo_file)
plot(logo, T)
cr_dim(logo)
tiny_logo <- cr_resize(logo, 200, 50)
cr_dim(tiny_logo)
plot(tiny_logo, T)
}
