% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-transform.R
\name{cr_rotate}
\alias{cr_rotate}
\title{Rotate}
\usage{
cr_rotate(cr, angle)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{angle}{rotation angle in radians}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Rotate
}
\examples{
canvas <- cr_new(800, 800, fill = 'royalblue')
cr_translate(canvas, 400, 400)
cr_rotate(canvas, pi/8)
cr_rect(canvas, x = 0, y = 0, w = 100, h = 100, hjust = 0.5, vjust = 0.5)
plot(canvas, T)
}
\seealso{
Other drawing transformation functions: 
\code{\link{cr_scale}()},
\code{\link{cr_transform}()},
\code{\link{cr_transform_reset}()},
\code{\link{cr_transform_set}()},
\code{\link{cr_translate}()}
}
\concept{drawing transformation functions}
