% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-transform.R
\name{cr_scale}
\alias{cr_scale}
\title{Scale}
\usage{
cr_scale(cr, x, y = x)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{x, y}{scale factors}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Scale
}
\examples{
canvas <- cr_new(800, 800, fill = 'royalblue')
cr_translate(canvas, 400, 400)
cr_scale(canvas, x = 1, y = 2)

# Draw a square in the transformed environment
cr_rect(canvas, x = 0, y = 0, w = 100, h = 100, hjust = 0.5, vjust = 0.5)

plot(canvas, T)
}
\seealso{
Other drawing transformation functions: 
\code{\link{cr_rotate}()},
\code{\link{cr_transform}()},
\code{\link{cr_transform_reset}()},
\code{\link{cr_transform_set}()},
\code{\link{cr_translate}()}
}
\concept{drawing transformation functions}
