% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{cr_to_array}
\alias{cr_to_array}
\title{Convert cairo raster to numeric array}
\usage{
cr_to_array(cr)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}
}
\value{
R array
}
\description{
Convert cairo raster to numeric array
}
\examples{
logo_file <- system.file("img/logo-rainbow.png", package = 'cara', mustWork = TRUE)
logo <- cr_read_png(logo_file)
arr <- cr_to_array(logo)
class(arr)
plot(as.raster(arr))
logo2 <- array_to_cr(arr)
plot(logo2, T)
}
\seealso{
Other image conversion functions: 
\code{\link{array_to_cr}()},
\code{\link{cr_to_nr}()},
\code{\link{cr_to_raster}()},
\code{\link{magick_to_cr}()},
\code{\link{matrix_to_cr}()},
\code{\link{nr_to_cr}()},
\code{\link{raster_to_cr}()}
}
\concept{image conversion functions}
