% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{cr_to_nr}
\alias{cr_to_nr}
\title{Convert \code{cairo raster} to \code{native raster}}
\usage{
cr_to_nr(cr, nr = NULL)
}
\arguments{
\item{cr}{\code{cairo raster}}

\item{nr}{Optional native raster to copy the cairo raster into. Must be same size.
Default: NULL means to create a new native raster object}
}
\value{
native raster
}
\description{
Convert \code{cairo raster} to \code{native raster}
}
\examples{
logo_file <- system.file("img/logo-rainbow.png", package = 'cara', mustWork = TRUE)
logo <- cr_read_png(logo_file)
nr <- cr_to_nr(logo)
class(nr)
logo2 <- nr_to_cr(nr)
plot(logo2, T)
}
\seealso{
Other image conversion functions: 
\code{\link{array_to_cr}()},
\code{\link{cr_to_array}()},
\code{\link{cr_to_raster}()},
\code{\link{magick_to_cr}()},
\code{\link{matrix_to_cr}()},
\code{\link{nr_to_cr}()},
\code{\link{raster_to_cr}()}
}
\concept{image conversion functions}
