% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{cr_to_raster}
\alias{cr_to_raster}
\title{Convert cairo raster to standard R raster}
\usage{
cr_to_raster(cr)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}
}
\value{
R raster
}
\description{
Convert cairo raster to standard R raster
}
\examples{
logo_file <- system.file("img/logo-rainbow.png", package = 'cara', mustWork = TRUE)
logo <- cr_read_png(logo_file)
ras <- cr_to_raster(logo)
class(ras)
logo2 <- raster_to_cr(ras)
plot(logo2, T)
}
\seealso{
Other image conversion functions: 
\code{\link{array_to_cr}()},
\code{\link{cr_to_array}()},
\code{\link{cr_to_nr}()},
\code{\link{magick_to_cr}()},
\code{\link{matrix_to_cr}()},
\code{\link{nr_to_cr}()},
\code{\link{raster_to_cr}()}
}
\concept{image conversion functions}
