% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-transform.R
\name{cr_transform_reset}
\alias{cr_transform_reset}
\title{Reset transformation}
\usage{
cr_transform_reset(cr)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Reset transformation
}
\examples{
canvas <- cr_new(800, 800, fill = 'royalblue')
cr_translate(canvas, 400, 400)
cr_rotate(canvas, pi/8)
cr_transform_reset(canvas)

cr_rect(canvas, x = 0, y = 0, w = 100, h = 100, hjust = 0.5, vjust = 0.5)
plot(canvas, T)
}
\seealso{
Other drawing transformation functions: 
\code{\link{cr_rotate}()},
\code{\link{cr_scale}()},
\code{\link{cr_transform}()},
\code{\link{cr_transform_set}()},
\code{\link{cr_translate}()}
}
\concept{drawing transformation functions}
