% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-transform.R
\name{cr_transform_set}
\alias{cr_transform_set}
\title{Set transformation matrix}
\usage{
cr_transform_set(cr, xx = 1, yx = 0, xy = 0, yy = 1, x0 = 0, y0 = 0)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{xx, yx, xy, yy, x0, y0}{params}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Clear the current transformation matrix and set it to these values.
}
\details{
\code{x_new = xx * x + xy * y + x0;}
\code{y_new = yx * x + yy * y + y0;}
}
\examples{
canvas <- cr_new(800, 800, fill = 'royalblue')

# Shear transformation
cr_transform_set(canvas, xy = 0.3)

cr_rect(canvas, x = 400, y = 400, w = 100, h = 100, hjust = 0.5, vjust = 0.5)
plot(canvas, T)
}
\seealso{
Other drawing transformation functions: 
\code{\link{cr_rotate}()},
\code{\link{cr_scale}()},
\code{\link{cr_transform}()},
\code{\link{cr_transform_reset}()},
\code{\link{cr_translate}()}
}
\concept{drawing transformation functions}
