% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-transform.R
\name{cr_translate}
\alias{cr_translate}
\title{Translate}
\usage{
cr_translate(cr, x, y)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{x, y}{translation offsets}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
Translate
}
\examples{
canvas <- cr_new(800, 800, fill = 'royalblue')
cr_translate(canvas, 400, 400)

# Draw a square at the origin of this transformed envionrment
cr_rect(canvas, x = 0, y = 0, w = 100, h = 100, hjust = 0.5, vjust = 0.5)

plot(canvas, T)
}
\seealso{
Other drawing transformation functions: 
\code{\link{cr_rotate}()},
\code{\link{cr_scale}()},
\code{\link{cr_transform}()},
\code{\link{cr_transform_reset}()},
\code{\link{cr_transform_set}()}
}
\concept{drawing transformation functions}
