% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr-png.R
\name{cr_write_png}
\alias{cr_write_png}
\title{Write cairo raster as PNG}
\usage{
cr_write_png(cr, filename)
}
\arguments{
\item{cr}{cairo raster}

\item{filename}{output filename}
}
\value{
None
}
\description{
Write cairo raster as PNG
}
\examples{
cr <- cr_new(800, 600, 'hotpink')
cr_write_png(cr, tempfile())
}
