% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pattern.R
\name{create_pattern_image}
\alias{create_pattern_image}
\title{Create a pattern specification using an image}
\usage{
create_pattern_image(cr, extend = "repeat", filter = NULL)
}
\arguments{
\item{cr}{\code{cairo raster} object as created by \code{\link{cr_new}()}}

\item{extend}{How will the pattern be handled outside the patterns natural area?
Default: "repeat".  Valid options: "none", "repeat", "reflect", or "pad"
\describe{
  \item{\code{none}}{pixels outside of the source pattern are fully transparent }
  \item{\code{repeat}}{the pattern is tiled by repeating}
  \item{\code{reflect}}{the pattern is tiled by reflecting at the edges}
  \item{\code{pad}}{pixels outside of the pattern copy the closest pixel from the source}
}}

\item{filter}{Set the filter to be used when rendering 
Default: NULL (use 'good' filter).  Valid values: 'fast', 'good', 
'best', 'nearest', 'bilinear'
\describe{
  \item{\code{fast}}{high performance filter with quality similar to 'nearest'}
  \item{\code{good}}{reasonable performance filter with quality similar to 'bilinear'}
  \item{\code{best}}{highest quality available. Performance may not be suitable for interactive use}
  \item{\code{nearest}}{Nearest-neighbour filtering}
  \item{\code{bilinear}}{Linear interpolation in two dimensions}
}}
}
\value{
pattern object
}
\description{
Patterns can be used as the \code{fill} argument to any of the drawing 
operations.
}
\examples{
logo_file <- system.file("img/logo-rainbow.png", package = 'cara', mustWork = TRUE)
logo   <- cr_read_png(logo_file)

canvas <- cr_new(1024, 726, fill = 'royalblue')


pat <- create_pattern_image(logo, extend = "repeat") |>
  cr_pattern_scale(0.2) |>
  cr_pattern_rotate(0.1 * pi)

cr_rect(canvas, 20, 20, 400, 400, fill = pat, color = 'black')

plot(canvas, T)
}
\seealso{
Other pattern creation functions: 
\code{\link{create_pattern_linear_gradient}()},
\code{\link{create_pattern_radial_gradient}()}
}
\concept{pattern creation functions}
