% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-pattern.R
\name{create_pattern_radial_gradient}
\alias{create_pattern_radial_gradient}
\title{Create a pattern specification for a radial gradient}
\usage{
create_pattern_radial_gradient(
  colors,
  stops,
  x1,
  y1,
  r1,
  x2,
  y2,
  r2,
  extend = "repeat",
  filter = NULL
)
}
\arguments{
\item{colors, stops}{color stops}

\item{x1, y1, r1, x2, y2, r2}{radial spec}

\item{extend}{How will the pattern be handled outside the patterns natural area?
Default: "repeat".  Valid options: "none", "repeat", "reflect", or "pad"
\describe{
  \item{\code{none}}{pixels outside of the source pattern are fully transparent }
  \item{\code{repeat}}{the pattern is tiled by repeating}
  \item{\code{reflect}}{the pattern is tiled by reflecting at the edges}
  \item{\code{pad}}{pixels outside of the pattern copy the closest pixel from the source}
}}

\item{filter}{Set the filter to be used when rendering 
Default: NULL (use 'good' filter).  Valid values: 'fast', 'good', 
'best', 'nearest', 'bilinear'
\describe{
  \item{\code{fast}}{high performance filter with quality similar to 'nearest'}
  \item{\code{good}}{reasonable performance filter with quality similar to 'bilinear'}
  \item{\code{best}}{highest quality available. Performance may not be suitable for interactive use}
  \item{\code{nearest}}{Nearest-neighbour filtering}
  \item{\code{bilinear}}{Linear interpolation in two dimensions}
}}
}
\value{
pattern object
}
\description{
Create a pattern specification for a radial gradient
}
\examples{
canvas <- cr_new(1024, 726, fill = 'royalblue')

pat <- create_pattern_radial_gradient(rainbow(10), seq(0, 1, length.out = 10), 
        x1 = 0, y1 = 0, r1 = 10, x2 = 1024, y2 = 726, r2 = 1000) 

cr_rect(canvas, 20, 20, 1000, 400, fill = pat)

plot(canvas, T)
}
\seealso{
Other pattern creation functions: 
\code{\link{create_pattern_image}()},
\code{\link{create_pattern_linear_gradient}()}
}
\concept{pattern creation functions}
