% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{magick_to_cr}
\alias{magick_to_cr}
\alias{cr_to_magick}
\title{Convert a 'magick' image to \code{cairo raster}}
\usage{
magick_to_cr(im, cr = NULL)

cr_to_magick(cr)
}
\arguments{
\item{im}{image from the \code{magick} package}

\item{cr}{optional cairo raster as a preallocation destination. If NULL (default), 
then a new cairo raster will be created}
}
\value{
\code{cairo raster}
}
\description{
Convert a 'magick' image to \code{cairo raster}
}
\examples{
\dontshow{if (interactive() && requireNamespace('magick', quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
im <- magick::logo
cr <- magick_to_cr(im)
plot(cr)
\dontshow{\}) # examplesIf}
}
\seealso{
Other image conversion functions: 
\code{\link{array_to_cr}()},
\code{\link{cr_to_array}()},
\code{\link{cr_to_nr}()},
\code{\link{cr_to_raster}()},
\code{\link{matrix_to_cr}()},
\code{\link{nr_to_cr}()},
\code{\link{raster_to_cr}()}
}
\concept{image conversion functions}
