% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{nr_to_cr}
\alias{nr_to_cr}
\title{Convert \code{native raster} to \code{cairo raster}}
\usage{
nr_to_cr(nr, cr = NULL)
}
\arguments{
\item{nr}{\code{native raster}}

\item{cr}{Optional cairo raster to copy the native raster into.  Must be same size.
Default: NULL means to create a new cairo raster object}
}
\value{
cairo raster
}
\description{
Convert \code{native raster} to \code{cairo raster}
}
\examples{
logo_file <- system.file("img/logo-rainbow.png", package = 'cara', mustWork = TRUE)
logo <- cr_read_png(logo_file)
nr <- cr_to_nr(logo)
class(nr)
logo2 <- nr_to_cr(nr)
plot(logo2, T)
}
\seealso{
Other image conversion functions: 
\code{\link{array_to_cr}()},
\code{\link{cr_to_array}()},
\code{\link{cr_to_nr}()},
\code{\link{cr_to_raster}()},
\code{\link{magick_to_cr}()},
\code{\link{matrix_to_cr}()},
\code{\link{raster_to_cr}()}
}
\concept{image conversion functions}
