% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-core.R
\name{plot.cairo_t}
\alias{plot.cairo_t}
\title{Plot a \code{cairo raster}}
\usage{
\method{plot}{cairo_t}(x, y, ...)
}
\arguments{
\item{x}{\code{cairo raster}}

\item{y}{any argument here will cause \code{grid::grid.newpage()} to 
be called prior to drawing the \code{cairo raster}}

\item{...}{other arguments passed to \code{grid::grid.raster()}}
}
\value{
None.
}
\description{
Plot a \code{cairo raster}
}
\examples{
cr <- cr_new(200, 100, 'hotpink')
plot(cr)

}
