% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cairo-core.R
\name{print.cairo_t}
\alias{print.cairo_t}
\title{Print method}
\usage{
\method{print}{cairo_t}(x, ...)
}
\arguments{
\item{x}{cairo_t object}

\item{...}{ignored}
}
\value{
None
}
\description{
Print method
}
\examples{
logo_file <- system.file("img/logo-rainbow.png", package = "cara")
logo <- cr_read_png(logo_file)
print(logo)
}
