% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{raster_to_cr}
\alias{raster_to_cr}
\title{Convert standard R raster to cairo raster}
\usage{
raster_to_cr(ras, cr = NULL)
}
\arguments{
\item{ras}{R raster}

\item{cr}{optional cairo raster as a preallocation destination. If NULL (default), 
then a new cairo raster will be created}
}
\value{
cairo raster
}
\description{
Convert standard R raster to cairo raster
}
\examples{
logo_file <- system.file("img/logo-rainbow.png", package = 'cara', mustWork = TRUE)
logo <- cr_read_png(logo_file)
ras <- cr_to_raster(logo)
class(ras)
logo2 <- raster_to_cr(ras)
plot(logo2, T)
}
\seealso{
Other image conversion functions: 
\code{\link{array_to_cr}()},
\code{\link{cr_to_array}()},
\code{\link{cr_to_nr}()},
\code{\link{cr_to_raster}()},
\code{\link{magick_to_cr}()},
\code{\link{matrix_to_cr}()},
\code{\link{nr_to_cr}()}
}
\concept{image conversion functions}
